Move HintProxy into separate file
